Move color changer function to its own doc
