+++ main.js
@@ -1,18 +1,9 @@
 import './fontawesome-free-5.9.0-web/css/all.min.css';
 import './assets/scss/styles.scss';
+import './scripts/color/color-changer';
 import './scripts/contrast';
 import './scripts/cssTemplate';
 import './scripts/icons/editIcons';
-
-// Set accent colour
-let colorChanger = document.querySelector('#colour_changer');
-
-colorChanger.addEventListener('input', () => {
-	document.documentElement.style.setProperty(
-		'--card-main-colour',
-		colorChanger.value,
-	);
-});
 
 // Code sample stuff
