+++ scripts/color/color-changer.js
@@ -1,0 +1,9 @@
+// Set accent colour
+let colorChanger = document.querySelector('#colour_changer');
+
+colorChanger.addEventListener('input', () => {
+	document.documentElement.style.setProperty(
+		'--card-main-colour',
+		colorChanger.value,
+	);
+});
